Update Tracking Library

+++ server/src/main/kotlin/suwayomi/tachidesk/manga/impl/track/tracker/mangaupdates/dto/ListItem.kt
@@ -17,5 +17,5 @@ data class ListItem(
 fun ListItem.copyTo(track: Track): Track =
     track.apply {
         this.status = listId ?: READING_LIST
-        this.last_chapter_read = this@copyTo.status?.chapter?.toFloat() ?: 0f
+        this.last_chapter_read = this@copyTo.status?.chapter?.toDouble() ?: 0.0
     }
